library-app: tests working with right database mocks

+++ misc-code/library-manager/app-user/tests/unit_test_db_conn.py
@@ -10,15 +10,16 @@ _ENGINE = create_engine(DATABASE_URL, echo=False)
 SessionLocal = orm.sessionmaker(bind=_ENGINE, autocommit=False, autoflush=False)
 
 
-def _reset_schema():
+def _reset_schema(reset_data=True):
     orm.session.close_all_sessions()
-    Base.metadata.drop_all(_ENGINE)
+    if reset_data:
+        Base.metadata.drop_all(_ENGINE)
     Base.metadata.create_all(_ENGINE)
 
 
 @contextmanager
-def get_db_session():
-    _reset_schema()
+def get_db_session(reset_data=True):
+    _reset_schema(reset_data)
     db_session = SessionLocal()
     try:
         yield db_session
